+++ internal/jwtverify/token_verifier_jwt.go
@@ -368,7 +368,7 @@ func (verifier *VerifierJWT) verifySignatureByJWK(token *jwt.Token, tokenVars ma
 	return verifier.jwksManager.verify(token, tokenVars)
 }
 
-func (verifier *VerifierJWT) VerifyConnectToken(t string) (ConnectToken, error) {
+func (verifier *VerifierJWT) VerifyConnectToken(t string, skipVerify bool) (ConnectToken, error) {
 	token, err := jwt.ParseNoVerify([]byte(t)) // Will be verified later.
 	if err != nil {
 		return ConnectToken{}, fmt.Errorf("%w: %v", ErrInvalidToken, err)
@@ -421,13 +421,15 @@ func (verifier *VerifierJWT) VerifyConnectToken(t string) (ConnectToken, error) 
 		}
 	}
 
-	if verifier.jwksManager != nil {
-		err = verifier.verifySignatureByJWK(token, tokenVars)
-	} else {
-		err = verifier.verifySignature(token)
-	}
-	if err != nil {
-		return ConnectToken{}, fmt.Errorf("%w: %v", ErrInvalidToken, err)
+	if !skipVerify {
+		if verifier.jwksManager != nil {
+			err = verifier.verifySignatureByJWK(token, tokenVars)
+		} else {
+			err = verifier.verifySignature(token)
+		}
+		if err != nil {
+			return ConnectToken{}, fmt.Errorf("%w: %v", ErrInvalidToken, err)
+		}
 	}
 
 	if claims.Channel != "" {
@@ -556,7 +558,7 @@ func (verifier *VerifierJWT) VerifyConnectToken(t string) (ConnectToken, error) 
 	return ct, nil
 }
 
-func (verifier *VerifierJWT) VerifySubscribeToken(t string) (SubscribeToken, error) {
+func (verifier *VerifierJWT) VerifySubscribeToken(t string, skipVerify bool) (SubscribeToken, error) {
 	token, err := jwt.ParseNoVerify([]byte(t)) // Will be verified later.
 	if err != nil {
 		return SubscribeToken{}, fmt.Errorf("%w: %v", ErrInvalidToken, err)
@@ -609,13 +611,15 @@ func (verifier *VerifierJWT) VerifySubscribeToken(t string) (SubscribeToken, err
 		}
 	}
 
-	if verifier.jwksManager != nil {
-		err = verifier.verifySignatureByJWK(token, tokenVars)
-	} else {
-		err = verifier.verifySignature(token)
-	}
-	if err != nil {
-		return SubscribeToken{}, fmt.Errorf("%w: %v", ErrInvalidToken, err)
+	if !skipVerify {
+		if verifier.jwksManager != nil {
+			err = verifier.verifySignatureByJWK(token, tokenVars)
+		} else {
+			err = verifier.verifySignature(token)
+		}
+		if err != nil {
+			return SubscribeToken{}, fmt.Errorf("%w: %v", ErrInvalidToken, err)
+		}
 	}
 
 	now := time.Now()